backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
python-multipart==0.0.6
werkzeug==2.3.7